change to bigint

+++ ITransitionTask3/Controllers/LcmController.cs
@@ -9,21 +9,21 @@ public class LcmController : ControllerBase
     [HttpGet]
     public IActionResult Get([FromQuery] string? x, [FromQuery] string? y)
     {
-        if (!int.TryParse(x, out int a) || !int.TryParse(y, out int b) || a <= 0 || b <= 0)
+        if (!long.TryParse(x, out long a) || !long.TryParse(y, out long b) || a <= 0 || b <= 0)
         {
             return Content("NaN", "text/plain");
         }
 
-        int gcd = Gcd(a, b);
+        long gcd = Gcd(a, b);
         long lcm = (long)a / gcd * b; // use long to avoid overflow
         return Content(lcm.ToString(), "text/plain");
     }
 
-    private int Gcd(int a, int b)
+    private long Gcd(long a, long b)
     {
         while (b != 0)
         {
-            int t = b;
+            long t = b;
             b = a % b;
             a = t;
         }
